requirements.txt
streamlit==1.31.0
pyyaml==6.0.1
fpdf==1.7.2
